py_requirements.txt
apache-airflow>=2.4.0
apache-airflow-providers-apache-spark==2.1.3
confluent-kafka>=1.8.2
apache-airflow-providers-apache-kafka
pendulum